requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_openai==0.1.6
langchain_text_splitters==0.0.1
numpy==1.26.4
pandas==2.2.1
plotly==5.22.0
scipy==1.13.0
sidrapy==0.1.4
streamlit==1.34.0
yfinance==0.2.38
python-bcb==0.2.0